+++ ArchiSteamFarm/MobileAuthenticator.cs
@@ -153,7 +153,7 @@ namespace ArchiSteamFarm {
 						return null;
 					}
 
-					if (!uint.TryParse(idString, out uint id) || (id == 0)) {
+					if (!ulong.TryParse(idString, out ulong id) || (id == 0)) {
 						Bot.ArchiLogger.LogNullError(nameof(id));
 						return null;
 					}
@@ -371,11 +371,11 @@ namespace ArchiSteamFarm {
 		}
 
 		internal sealed class Confirmation {
-			internal readonly uint ID;
+			internal readonly ulong ID;
 			internal readonly ulong Key;
 			internal readonly Steam.ConfirmationDetails.EType Type;
 
-			internal Confirmation(uint id, ulong key, Steam.ConfirmationDetails.EType type) {
+			internal Confirmation(ulong id, ulong key, Steam.ConfirmationDetails.EType type) {
 				if ((id == 0) || (key == 0) || (type == Steam.ConfirmationDetails.EType.Unknown)) {
 					throw new ArgumentNullException(nameof(id) + " || " + nameof(key) + " || " + nameof(type));
 				}
